refactor: change arhitecture oof folders
