+++ src/components/seeds/amount-panel/AmountPanel.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import { MagicContext } from '../../context/MagicProvider';
-import Button from '../button'
+import { MagicContext } from '../../../context/MagicProvider';
+import Button from '../../reuse-components/button'
 import './AmountPanel.scss';
 
 const AmountPanel = () => {
