+++ src/components/seeds/amount-panel/AmountPanel.scss
