+++ src/components/seeds/amount-panel/index.js
