+++ src/components/seeds/card-list/Modal.scss
