+++ src/components/seeds/card/Card.scss
