+++ src/components/seeds/card/index.js
