+++ src/components/seeds/page-counter/PageCounter.js
