+++ src/components/seeds/page-counter/PageCounter.scss
