+++ src/components/seeds/page-counter/index.js
